refactor: split controllers
